Events defined Network defining simplified Handling active connections simplified Gathering statistics moved to class Network from class Simulator Memory management added
Event classes created
Simulator class created
Logger class created
Network defining done
Establishing connections done
Checking for internal and external blocks done

+++ include/Link.h
@@ -13,7 +13,7 @@ public:
     uint64_t sourceNode;
     uint64_t destinationNode;
 
-    Link(int sourceNode, int destinationNode);
+    Link(uint64_t sourceNode, uint64_t destinationNode);
 
     void reserveFSUs(uint16_t firstFSU, uint16_t numberOfFSUs);
